ml/deeplearning/requirements.txt
bz2file==0.98
cython==0.29.33
numpy==1.24.2
flask==2.2.3
gensim==4.3.1
h5py==3.8.0
jupyter==1.0.0
matplotlib==3.7.1
pandas==1.5.3
pillow==9.4.0
pip==23.0.1
pytest==7.2.2
pyhamcrest==2.0.4
scikit-learn==1.2.2
scipy==1.10.1
sklearn==0.0.post1
seaborn==0.12.2
setuptools==67.3.2
twisted==22.10.0
tensorflow-macos==2.11.0
